REDS_reward_learning/pyproject.toml
[tool.poetry]
name = "bpref-v2"
version = "0.1.0"
description = ""
authors = ["John Smith <john@example.com>"]
readme = "README.md"
packages = [{include = "bpref_v2"}]

[tool.poetry.dependencies]
python = ">3.10"
numpy = ">=1.20.2"
scipy = ">=1.6.0"
absl-py = ">=0.12.0"
gym = ">=0.18.0"
gdown = ">=3.12.2"
tqdm = ">=4.60.0"
flax = ">=0.3.5"
ml-collections = ">=0.1.0"
optax = ">=0.0.6"
tensorboardx = "^2.6.1"
tensorflow-probability = ">=0.14.1"
imageio = ">=2.9.0"
imageio-ffmpeg = ">=0.4.3"
pandas = "^2.0.3"
d4rl = {git = "https://github.com/Farama-Foundation/D4RL.git"}
hydra = "^2.5"
opencv-python = "^4.8.0.74"
hydra-core = "^1.3.2"
pre-commit = "^2.12.0"
black = "^23.3.0"
pytest = "^7.4.0"
isort = "^5.12.0"
transformers = "^4.30.2"
rich = "^13.4.2"
ujson = "^5.8.0"
wandb = "^0.15.5"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.isort]
# https://pycqa.github.io/isort/docs/configuration/black_compatibility/
profile = "black"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
